requirements.txt
asgiref
certifi
charset-normalizer
coreapi
coreschema
Django
django-filter
djangorestframework
drf-yasg
idna
inflection
itypes
Jinja2
MarkupSafe
packaging
Pillow
psycopg2-binary
pyparsing
python-decouple
pytz
requests
ruamel.yaml
ruamel.yaml.clib
sqlparse
uritemplate
urllib3
coreapi